refactor: move MatchingRoundPhaseView to web layer and unify phase type across responses
